# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Event Access API
  description: API for event registration, QR code generation, and access control
  version: 1.0.0
  contact:
    name: Event Access Team
    email: support@eventaccess.com
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: /api/v1
    description: Production server
...
paths:
...
  /admin/participants/{id}/accept:
    post:
      tags:
        - Admin
      summary: Accept a participant
      description: Accept a participant's registration
      operationId: acceptParticipant
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Participant accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Participant not found
        '401':
          description: Unauthorized
  /admin/participants/{id}/reject:
    post:
      tags:
        - Admin
      summary: Reject a participant
      description: Reject a participant's registration
      operationId: rejectParticipant
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Participant rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Participant not found
        '401':
          description: Unauthorized
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
...